Monkey path a problem in gdb's python integration.
The gdb-for-python somehow gets confused and thinks that PyUnicodeObject is
a struct, instead of a typedef. This is probably related to how the shared
object we're building is exporting symbols in some funny way, but either way
gdb blows up whenever we do anything after loading typed_python, even if we
remove all of the content from it. So, probably a compiler problem. This
hack ensures that it will work, instead of giving us

 <class 'RuntimeError'> Type does not have a target.:

everywhere in the gdb stacktraces, which is what happens when you try to
ask for .target() on the PyUnicodeObject 'struct' (without importing the module
it's a typedef)