Forward types working.
I decided to go with the syntax

   X = SomeTypeExpression(lambda: X)

with the idea that 'lambda: X' means "please evaluate this
type expression when we need the type to be concrete, not now
when I'm defining types." This is in line with how mutually
recursive classes in python find each other (through their
respective closures).

We have the additional requirement that the relationship between
types never changes, and we evaluate the lambda function once
and only once, when we first attempt to do something concrete
with the type.

I also implemented an early working version of the compiler here.
This version produces code that works for simple integer calculations,
but the abstractions for managing code and entrypoints don't line up
cleanly with the new type model.

In the future, we're not going to compile individual entrypoints and
invoke them directly (which is how this version works). We're
going to either JIT, interpret, or precompile. In all cases, entrypoints
to compiled code will be hung off of typed_python Function objects
and that'll get managed by an internal framework. This version
still has the old model building up 'native function pointer' objects
and invoking them with ctypes.