Implement object serialization when we have a type <-> name lookup table.
This is intended to substitute for pickle, but it allows us to more easily control
how types are looked up from strings. This means we can coherently serialize from
one codebase and deserialize to another within a single process, which pickle doesn't
give us enough control to do. We also can pass back and forth between native and
pythonic serialization within the same serialization buffer.

We still need to implement something to handle lambdas and closures within a
codebase context.