Add a temporary shim to allow us to see exceptions we're throwing in native code.
Right now, we have to stash the exception as a const char* in a thread-local
temporary and then read it in the invoking code.

Over the long run, we will model exception types and try/catch
semantics correctly, and the dispatch code will be reponsible for
catching any exceptions coming out of native code and translating them.