Rework how we handle intermediate expressions to use 'scopes'.
In this model, we have an expression context for each expression
during which temporaries will be kept alive (for instance, during
a single statement).

The context allows us to push variable definitions that have side
effects onto a stack, so that we don't have to worry about subexpressions
that have side-effects showing up more than once. This is close
to the same model that llvmlite uses - you the context is like a 'builder'
allowing us to build up a sequence of expressions and using Context Managers
to handle flow control.

Under this model, it should be much easier to ensure we track
refcounts appropriately.

As part of this change, we also infer the complete set of variables
we have assigned to and are careful to track whether they have
been initialized at the start of the function.