Output type of functions is now inferred.
Also, native functions have a crude type-based dispatch mechanism. This
should eventually be stored as a tree, so that we can efficently tell if
a given set of python values matches the specialization. We should also
be careful about ordering of specializations, since if a less-specialized
function gets compiled before a more-specialized one, it will mask the
more specialized one because our dispatch is a simple linear search through
the various specializations, taking the first one that works.

But for now, this is enough.