Convert groups of functions simultaneously.
This is a very rough cut at a good type-inference algorithm.
The idea is that we have a set of functions we're converting based
on which python function they are and what their input types are.
Output types are unknown in the case of mutually recursive functions,
so we repeatedly regenerate code until the types stabilize. As it
stands, we regenerate functions many times agressively, because
we're not tracking which functions are dirtied when other functions
change their return types. We are also not going to catch it
(yet) when we have an infinite set of types - we'll just keep
generating code until the user breaks the program. Eventually
we'll want a way for users to control this, with a bailout to
the interpreter if we can't find anything good.