Implement ListOf.
As part of this, change how tuples are implemented, because we need
both tuples and lists (which share an underlying implementation) to
be able to be resized and have that track with all references to the
object. This wasn't an issue with tuples because they're immutable,
but lists can be resized, which means that the object on the heap
needs to be stable, so it has to have a pointer internally that
we can resize as needed.