Create a 'TypeFunction' abstraction. Analagous to templates in C++.
This makes it possible to talk about parameterized types. For instance,
write

        @TypeFunction
        def List(T):
            return Alternative("List",
                Node={"head": T, "tail": List(T)},
                Empty={}
                )

and you have a parameterized ml-style list. Results of type functions
are memoized, and the codebase/serialization/compilation framework
knows about them. This is the primary way to talk about sets of
related types, or anonymous types, because we can give them
a formal identity (the type function that produced them, which
should be a module-level object, and the arguments to said function).