Build a shared object, not llvm bitcode, when compiling a whole codebase.
This lets us load the resulting module as a .so library, without having
to respecialize the bytecode for the target each time. This can be MUCH faster,
but requires that the binary is only used on the same target, or targets
that are equivalent.

We have to link the llvmlite output (which is basically a .o file) into
a shared library. We're using `ld` for that, which should exist on most
linux systems. But it would be better to use the pip-installable package
`llvm` (as opposed to `llvmlite`) to do this as its cross platform and
the current solution will only work on linux.