Release the GIL when we enter compiled code.
This change means that compiled code can segfault if
we peform multithreaded non-sychronized access to list,
tuple, or dict items, or if we resize lists and dicts in
a nonsynchronized way.

This is equivalent to how multithreaded access to datastructures in
C++ works: you can read objects from multiple threads without
synchronization, but without it you cannot modify a datastructure.
But it's not how python works, so it could create unexepected crashes
if we compile compliant python code.

It would be good to later add a mechanism to allow us to check that
compiled code is accessing values in a compliant way by generating
a slower-but-safer form of code. In the short run, assume that
race conditions on list/dict items and list resizing operations
could segfault.