Fix some issues in ServiceManager tests.
We now use the gbRamUsed and coresUsed in a Service explicitly when we allocate
the process. During tests, we need to ensure we don't hit caps associated with
the machine itself. We should put an explicit amount of resources in the test harness
so we can actually test this behavior without depending on the state of the machine.