Disallow lines above 149 chars & Lint fixups
We eventually want enforce shorter line length but for now we
upper-bound it at a length where we didn't have may violations, so that
it's easy to enforce.

Also, remove all file-level noqa (there was just one), and replace with
a "per-file-ignores" directive for flake, which is more fine-grained
because it only excludes one specific check (we don't warn about unused
imports in __init__.py files because those define the module's API).

Also, replace all line noqa with qualified "noqa: Exxx"