Adjust semantics of cell.findChildrenByTag
When we pass the stopSearchingAtFoundTag we stop the recursive search at
nodes that match our query tag. Previously, we would stop the search at
any tagged node, which would cause tests to break when we tagged nodes
higher in the path from the root node. This mechanism of finding tagged
nodes should be fairly resilient to structural changes to the frontend
layout, which that behavior was not supporting, so we changed it.