Implement a more flexible encoding where we can add and remove fields from named tuples.
This new format is inspired by google protobuf, but matches our internal object model
more closely.  Objects are encoded with a field number and wire type encoded in a varint,
followed by their data. The structure of that data is determined by the wire type, so
that we can parse the structure of the message without knowing the type being serialized.

Consumers can then decide how to handle data contained in various field numbers. If we add
a field to a named tuple, or a new type of alternative, old code can handle it without
breaking by deserializing to a default value.

We don't currently have a way of setting the field numbers, but we can add one later
without violating the binary protocol.

See types_serialization_format_test.py for a more complete description of the protocol.