Fix a nasty bug in AlternativeType.
Recall that the evaluation order of subexpression expressions in C++ is unspecified
in many cases, and that if you write

	container[key] = container.size()

you may find that the 'container.size()' could be evaluated before or after 'container[key]'.
The version of gcc on Travis evaluates the key lookup first, so container.size is 1 larger
than the version of gcc I run with locally.