Ensure we can convert numpy.int64 and numpy.float64 into int/float in ODB and typed python.
As part of this change, work through some issues in how we convert types from one to another.
In general, we should be attempting to not convert if we don't need to. So for function dispatch,
if we call with an int, and both an int and a float option are available, we should pick the int
option, but if only a float option is available, we should call that.

This requires that we're much more careful about the 'isExplicit' argument floating around
in the system, since we want to be able to tell whether conversion will succeed (otherise,
it's too slow), and we may need multiple passes through a set of options to get
conversion to have the right semantics.