Take the new ForwardTypes system over the line.
This commit does a little cleanup and fixes the type-resolution process. We explicitly
track, for each Forward, all the types that need it, and for each type, all the forwards
that we're waiting on for reslotion. When a forward resolves, we update these data structures,
and so we can tell, when we resolve a forward type, whether it is resolving a cycle, because
then its new target will depend on it (and only it) as a forward.

Once we see a type cycle get resolved, we walk all the types and update them. We're also careful
to ensure that we don't allow a containment cycle, which would produce an object with
infinite bytecount, and to ensure that if we detect a reference cycle that we set the
object's name appropriately (to whatever was in the forward) to avoid a name recursion.