Fix a bug causing too many revision conflicts with full consistency. Insist on full consistency.
We had an old notion of consistency that's outmodeded - everything is 'full' now. In particular,
if you update fields that participate in an index, a write to that field is also an implicit read,
and if you lookup objects using an index, you have a read on the index. However, if you write
to an unindexed value without reading it, you donn't have a dependency on the prior value.

This means that you can effect the old style of consistency with more explicit semantics.

If you want to update a value without caring about its current state, simply write into
it without reading (this is equivalent to turning off read consistency). If you want to
lazily depend on the values of an index, read them in a view, then open a transaction
and loop over them, checking whether they exist. These patterns effect the existing
semantics, but force the race conditions inherent in them to the code making the transactions,
which is a good thing, because they are clearer at that layer.