Fix a super-nasty bug causing memory corruption during multithreading.
We release the GIL during deserialization. The GIL is the lock that
protects all of our view internals. When we deserialize wire-data to
python objects for a concrete type, which happens deep inside of
the VersionedObjects code, we can cause other threads to read from
the same VersionedObjects code at the same time, which can
reallocate the hashtables underneath us.

We should figure out something better. But it's not obvious what,
because if we are not careful we'll produce deadlocks. Probably it
would be best to have the view internals get protected entirely
by their own lock, and have callers do the deserialization at the
python-interop layer.