Remove a lock causing us to deadlock when we load Function types from multiple threads.
We should figure out how not to need to hit the interpreter when we call into
mirrorTypeInformationIntoPyType because it makes this code very hard to reason
about.  We hit the interpreter to build FunctionOverload objects, and this can
cause other threads to run and load the same object we're already building!