Ensure a cell doesn't unsubscribe randomly.
We have a bug where cells in the UX get 'stuck' because they unsubcribe
from changes propagating from object_database. This fix is topical.
Really, we should produce a more explicit lifecycle for cells. Cells exist
with an identity for a fixed lifetime. The reuse of cells is intended
to allow cells to be treated like factory functions (off, then on again).
This confuses the implementation because we have to maintain state
indicating whether a python cell object was re-used.

This commit deals with an annoying bug, but we could make the
model more rigorous by explicitly tracking which cells were
created and destroyed and removing that responsibility from the cell
itself.