Tuple instances can be iterated and []'d in compiled code.
We are careful to track the type information precisely in both cases
if at all possible. For instance, if `t` is of type `Tuple(int, float)`,
we know that `t[0]` is an int, and `t[1]` is a float because we can
see the exact integer constants contained within. Similarly,
iterating over a Tuple will loop-unroll because we know exactly
how big each tuple is.