Rework how we do type-conversion so that we can mimic typed-python's multi-pass semantics.
Especially when converting to something like a OneOf, we need to ensure we first check if there
are any non-explicit conversions before we check explicit conversions. This means we need a way
for compiled code to communicate whether conversion has happened to the calling context.
So we rework the conversion code so that it's phrased as a function

    bool convert(target_type* target, source_type& src)

where the bool indicates whether we've converted or not.