Allow the compiler to convert 'isinstance'.
This isn't smart enough to handle cases where the type is not known at compile time.
We eventually need to make sure we can convert any object in the compiler into a python
object, so we can use that functionality. But for the moment, this should allow us to
compile standard strongly-typed code against OneOf and Alternatives.