Move recently added terms in python_ast.Statement and Expression to the end of each Alternative .
Right now, we use the position of the ConcreteAlternative in the list
of alternatives to encode it when we serialize it. We intend to add an ability to
control how each alternative is serialized, so that when we add new alternatives
we can put them wherever we want in the list and give them new tags. But for right
now we don't have that functionality. As a result, the prior changes accidentally break
old serialized code, and this change un-breaks those serializations.