Be more careful about the order in which we compile functions. Refs #120.
This doesn't solve the problem of not compiling functions we no longer
need, but it does solve the problem of compiling functions many times in a row.
If we were willing to do the conversion recursively, then we could fix that as well,
and evaluate most functions exactly once.

Still, on larger compilations, we are no longer the bottleneck - llvm
is.