Split our view of unicode props up by python 3.6/7.
Apparently some characters are considered alphanumeric in 3.6 that
are not in 3.7. We want this to be fast on our side, so we have a lookup
table in code, and that needs to change based on python version.

Also, fix some minor differences that show up in the arithmetic
compilation tests between 3.7 and 3.6.