Ensure we can convert Class instances to object. Closes #252.
We weren't calling the class_wrapper's super() in convert, which has
the pathway that allows T1.convert_to_type(T2) to defer to 'T2.convert_to_self(T1)'.

I also added a check that if we claim we could convert, then we must
return a True constant out of the conversion code to try to catch this kind
of mistake.