PythonObjectOfType now held as a PyObject** instead of a PyObject*.
This means we have a separate atomic refcount we can use to track
the lifetime of the object, and avoid inc/decreffing PyObject when
we move objects around on our compiled stack.

This is in preparation for #238. We want to be able to interact
with some python objects where we know the datastructure and
don't need to hit the gil to interact with it.