Add a notion of a 'Module'.
Forward types are a little verbose to describe. This change
creates 'Module', which lets you describe mutually recursive
types more naturally. Accessing a member of a Module defines
a forward type for that member. This makes it possible to
refer to forward types without needing to place them
in a (python) module-level scope as explicitly declared
forward types.

E.g.

    m = Module("M")
    m.Json = OneOf(None, int, float, TupleOf(m.Json), ConstDict(str, m.Json))

as opposed to

    Json = Forward("Json")
    Json = Json.define(
        OneOf(None, int, float, TupleOf(Json), ConstDict(str, Json))
    )

which is a little more obtuse.