Ensure we get the right forward type out of the type memo. Fixes #235.
Before, if you wrote TupleOf(T) where T was a forward, and you then
defined the forward to F, TupleOf(F) wouldn't resolve to the same
thing as TupleOf(T). This creates huge problems because we expect
to be able to uniquely identify TupleOf/ListOf types in this way.

This is mostly an artifact of how we're wrapping the underlying
Type* objects. We can't walk over all the PyType objects and
fix them up when we define a forward, so cases where we have
TupleOf(F) and we resolve F to something like 'int' will always
be different objects than the core TupleOf(int). Forwards should
really only be used when we have actual type cycles.