Make ConstDict/Dict '.keys/values/items' look more compliant. Fixes #263.
We actually ought to make these things their own types. As it stands
we're using some flags in the PyInstance object to indicate that an instance
of a ConstDict is acting as 'keys'. As a result, we still have the bug
that type(ConstDict(int, int).keys()) is 'ConstDict(int, int)'.