Pass 'isStr' into Type::repr instead of holding in ReprAccumulator.
Whether a call is '__str__' or '__repr__' actually depends on
not only the calling context but also the state machine of 'repr'
itself. When we call 'str([a, b])', Python calls 'repr' on
'a' and 'b'. So instead of attaching the 'reprness' of the call
to the accumulator, we pass it down the call stack as part of the
call.