Compile dict.clear and dict.update, and fix Dict.clear in the interpreter. Refs #269.
In interpreter code, we were assuming that calling 'Dict.remove_key' didn't compress
the item table, which led to an iterator getting invalidated and breaking everything.

We also somehow were assuming that calling 'std::memset' on the dict_internals that
are supposed to be int32_t would work, even though it sets individual characters.
I'm mystified that this didn't segfault. I guess where we were using it it just
never mattered.