Enable cross-module inlining.
We used to have a problem where if we defined a function in @Entrypoint,
it couldn't be inlined by later users because we'd always refer to it
as an external reference. This change allows small functions to be redefined
inside of subsequent modules (at the llvm level), so that their definitions
are available to LLVM for inlining.