Introduce 'masquerading' types to the compiler. Fixes #241.
Functions like str.split return typed results (ListOf(str) in this case)
but need to _look_ like untyped results (list) to match the behavior of the
interpreter. We don't want to lose the typing information, but we must be
careful, because the resulting list is mutable, and users could place
anything they want inside the list, which would then mean that the list
was no longer exclusively made of strings.

To handle this problem, we introduce 'masquerading' types, which are
Wrapper objects that declare that they are masquerading as an untyped object.

These types attempt to look like the untyped type (e.g. 'list') but retain
the typing information. We make the assumption that if the untyped type is
mutable, then there is no more than one reference to the object at any point
of execution, and as soon as the object gets assigned to a variable, we
convert it to its untyped form.

As part of this change, correct the compiled signature of 'str.split'
which was not compliant with the interpreted form.