Implement closures in compiled code. Refs #236.
This change implements both passing closures into compiled code and
creating closures from within compiled code and passing them to other
functions.

It's not a perfect implementation - specifically, we leak the closures
of mutually recursive functions, because we don't have a way of breaking
the refcount cycle yet. In practice, this won't come up much but it's
something we'll need to hand.

To implement this, we modified 'Function' to allow it to contain a
reference to a closure object, and a mapping that dictates how a given
closure variable can be found vis-a-vis the closure. The closure object
is either the empty tuple (in the case of stateless mutually recursive
functions), or a Class instance with one member for each variable
bound in the closure.