Build a simple matrix/array library that links to fortran for matrix inversion / multiplication routines.
Because this gets compiled, for many operations on small matrices it is much faster than numpy because
there is no python overhead. It also does not touch the GIL, which means that it paralellizes well.

This code is here mostly to provide a concrete example of using typed_python to do something interesting,
and to provide a richer set of client code for testing purposes. But you can also use it to do stuff.