Implement 'HeldClass' semantics in the interpreter.
In this model, a container type (class, list, etc.) can "hold" a
class internally, meaning that instead of holding instances of the
class via a pointer with a refcount, they hold it within their
own elements inline. Python variables themselves can never be a
Held(C) - they will instead be RefTo(C), which is a _pointer_ to
the held element. This is inherently unsafe, but has the benefit
that it's (a) faster and (b) uses much less memory.

RefTo(C) is intended to behave mostly like a C instance, except
that you must ensure the storage that's backing it doesn't change.
You can copy a RefTo(C) back onto the heap as a regular C,
and you can get a RefTo(C) from any 'C' using the builtin
'refTo' function. This ensures you can safely move objects back
and forth between standard python semantics (with refcounts
everywhere) and reference semantics which are faster but
dangerous.

This is not implemented in the compiler yet.