Implement compiled 'map' and 'reduce' functions.
These are primarily used to work with tuples where we don't
know the input size. If you have a 'Tuple(t1, t2, ...)', you
can call 'map(f, t)' to apply a function to each interior element
of the tuple in a way that retains the type information and which
the compiler can optimize. If you had tried to build up an
untyped tuple instead, you'd hit the interpreter and everything
would be very slow.