Fix compilation and serialization of deserialized closures.
The issue is that we depend on the __code__'s file and line number to be local
on the machine to get access to the AST. This doesn't work when we have deserialized
the code from another machine.  So we have a cache that keeps track of the functions
that it creates by evaluating an ast.

In general, it would be nice to get a better way to convert between python bytecode
and the py-ast than what we do right now which is to convert to text. We're maintaining
a reverse map for all the objects we manually do this to, but it's not very clean.