Fix some try/catch control flow bugs.
In particular, we weren't distinguishing (in native_ast) between TryCatch
where the expression type is always 'void' and cases where the catch clauses
always rethrow (which we use in stack unwinding). Technically, we could
have a notion of a 'typed' try/catch expression, where the return values from
both the try and and the 'catch' clauses were of the same type. But we don't
have such an idea in python since try/except is always at the statement level,
and statements are not expressions.  So, in this case, it is sufficient
to split our try/catch handling into two cases - real Try/Catch as a statement,
where even if the try doesn't ever return, the catch very well might, and the
expression value is always Void, and stack unwinds, where if the 'try' portion
has no result, neither does the whole expression.

This is enough to ensure that our blocks are systematically terminated.
It would be good to also determine if the 'catch' handler doesn't return,
in which case the entire expression has no return value, but practically
speaking this is kind of irrelevant because the typing layer above
will already know that, and the conversion layer below (llvm) will
figure it out, so it won't be slower.