Fix a bug causing functions to not rebuild.
We iterate through the graph of functions and rebuild functions
when their call targets output types change. This is how we do
type inference in the presence of mutual recursion: if 'f' calls 'g'
and 'g' calls 'f', they both get initialized with no return type,
and then as we learn the output types of 'f' and 'g' we recalculate
each of them until the type graph is stable.

If we don't recalculate, we can end up with a wrong notion of
the output type and everything falls apart. In particular, we had
a bug where we failed to register the dependency except on new
functions. Reproing was actually somewhat tricky.