Staticmethods that have forward refs can still be instantiated. Fixes #320.
They just can't be called until their forwards are resolved.

Although really, it would be better if we figured out how to
wire into our Class type's get/set descriptors so we could produce
the staticmethods on demand instead of jamming them in the type's
dict object, because then we could check this when they were being
used. Although even that has problems because it's reasonable
to hold onto a closureless function object that has unresolved
forwards. You just shouldn't be able to call it.