Get rid of BoundCompiledMethod and make BoundMethod just work for everything.
We were having a problem with OneOf and BoundCompiledMethod because it didn't have
a proper typeRepresentation, and we need for typeRepresentation to be 1 <--> 1 with
type wrapper (at least at the moment) for the main framework to work correctly.

This change allows the BoundMethod type object to handle any kind of first arg
type, even if the interpreter doesn't (at this point) know how to do the dispatch
correctly, since all we're really doing is hijacking the Type object to hold onto
the first arg type and the attribute name.