Add a facility for placing function boundaries around repeated code (like oneof stuff).
Before this, we were explicitly inlining every one-of expansion for things like
call-ing, getattr-ing, and binary comparisons. This can get very expensive when
you have expressions involving two oneofs because you get all the pairs of types,
and then you repeat these in every location. It's much more efficient to create a
function to do the operation and then let llvm choose whether to inline it.