Get rid of Int64, Float64, Bytes, String, NoneType, Bool, as they are useless and confusing.
There's no reason to have separate models for builtin python types in typed_python. It just
makes code confusing because you write Tuple(int), get the element type out, and it's Int64.
But you can never construct an Int64!

In theory, we could retain Int64, so that we can be explicit in the interpreter about
mimicking the compiler semantics, but for 99% of code, we should just have 'int' and
that's it.