Redo how we do serialization and codebases.
Our first cut at this supported "multiple codebases at once", meaning
you could have two versions of a codebase for a module and have both of
them resident at the same time. This supported the object_database having
multiple versions of a codebase resident in the ActiveWebService. However,
it really overcomplicates things, because lots of python infrastructure
assumes that there's only one version of a given module by name (and
rightfully so).  This change rips that out, and replaces it with
a simpler model where a given foreign codebase can be instantiated
once per process, and it's a one-way operation.

The upshot is that SerializationContext now looks much more like
pickle in how it finds its dependencies - modules and classes
are looked up by name. The side benefit of this is that we now
have formal names for everything that are directly defined by
the code, solving a long-standing problem in object_database where
small changes to the code could 'rename' objects. It also
removes the need for us to import all modules in a codebase
before serializing from it, which has also been a problem in
other contexts.

As part of this change, Alternative attempts to figure out
the name of the module it's being defined within by looking
at the calling function's globals. A bit of magic, but
somewhat necessary.