Always check for __typed_python_module__ before __module__ when serializing.
Python doesn't actually allow you to modify a type's tp_dict after it's
created, because it reads some things out of it (apparently __module__ is one)
when the type is first created. So we can't update it after.

Instead, we set __typed_python_module__ instead of __module__ when we need
to communicate the module location of a type after the fact.