Make sure we can serialize Function objects at the module-level using names.
Unfortunately, because of our approach it's not always possible to correctly
override __qualname__ on our various types, much as we can't always correctly
affect __module__. Probably, with a little more massaging, this could work.

In the meantime, __typed_python_qualname__ and __typed_python_module__ are
our indication of our desire to override the non-typed_python forms of those
variables.