Make sure deserialization uses our existing type-hashes so we don't get multiple copies of the same type.
There is still an issue where we could be deserializing a mutually recursive group hierarchy in different
threads and end up with two copies of the same types linked up slightly differently. We aren't supposed
to depend on _python_ type identity per-se - we should be using the typehash. but eventually that might be
a problem, so something to keep in mind.