NotCompiled functions don't capture closures as globals.
We had been assuming that if you mark a function NotCompiled that
it was at module scope and so it was safe to embed its closure in its type.
This creates a huge leak if you repeatedly create NotCompiled functions
that capture surrounding variables with closures, because those variables
get encoded in the Type object, which we never release.

This change moves the closure back into the actual function closure,
and makes sure we can correctly compile against the resulting object.