Fix some bugs, improve string split perf, and make masquerades more usable.
I removed the pattern of 'converting the mutable masquerade to the untyped form'.
The problem is that if you have something like

	x = aString.split("\n")

then we model 'x' as a ListOf(str) in the compiler. When we return to the interpreter
then it goes back to a normal list. In the compiler, however, we want to keep the typed
form for performance reasons, but we can't guarantee that user code won't insert
a non-string into the list, violating the type assignment.

I decided that rather than trying to do some complex reference tracing thing, that
even if it works will still be brittle, it would be better to simply acknowledge
that the compiler will have different behavior from the interpreter in this case,
since for the most part it's not a great idea to be modifying the results of
string splits, or *args/**kwargs, and converting to the untyped form makes the
compiler kind of useless.

As part of this change, also switch to using PyMem_Alloc instead of malloc,
which appears to have some perf benefits.  Eventually we'll need to look
more carefully at the memory model - things that are guaranteed to be threadlocal,
for instance, could be allocated on a different heap. But for now, this is good enough.