Fix a bug in compiled min/max. binops can return non-bools, so we have to cast them.
Also, min/max was trying to unfold a complete tree expression for all the possible
comparison paths you could have for small numbers of arguments. Unfortunately, the
code size is actually N!, not N^2, and so it's prohibitive to compile. Instead,
if we have unlike types, we just need to convert everything to the union type.