Fix a multithreading bug where the compiler could get an unpopulated Function object.
We memoize the python PyTypeObject for each Type object in PyInstance. When we
are first creating it, we put it in the memo, and then fill out its details.
Function objects are complex in this regard, because they refer to themselves
in their "overloads" object, and they have to call into the python interpreter
to build this object. This means that if the interpreter switches to
another thread, that thread could get the memoized PyTypeObject before it
is filled out with details.

In the test I added here, we were failing because periodically, one of the two
threads would allow the other to run after PyInstance had memoized the object
but before 'mirrorTypeInformationIntoPyTypeConcrete' was finished. This
meant that the 'overloads' object wasn't populated, and so the test fails.
This is a behavior we observed in the wild.

To fix it, I modified the code to build the relevant objects without
hitting the interpreter at all. This requires that I explicitly call
tp_new for the class and fill out the dict manually.