Fix a bug serializing functions.
We were including the globals in the function's factory args,
which can create a circular dependency when we deserialize the
function. Instead, we need to be constructing a blank function
with the relevant code object and then filling out its
globals and closure. That way the function object exists
and is populated in the memo when we go to deserialize the
globals dict and closure.