Fix a bug where serializing a function twice in a row changed its hash.
We rely on being able to serialize a function without file or line information
in order to hash it (so that hashes are stable to moving code around).

We were stashing ast objects in python_ast._originalAstCache, and in some cases
stashing the version of the code that had the line info removed. We were also
then using the version without the line info removed if we found it in the
cache.

This change ensures we are more careful about this.